Restructured directory content

+++ .eslintrc.js
@@ -2,7 +2,7 @@ module.exports = {
   "extends" : "eslint:recommended",
   "rules" : {
     "semi" : ["error", "always"],
-    "quotes" : ["error", "double"],
+    "quotes" : ["error", "single"],
     "indent" : [2, 2, {"SwitchCase" : 1}],
     "no-console": 0,
     "no-var" : 1
